repositioning units number output, beginning persisting
where units output is displayed has been moved up and will be visible as soon as findTotalUnits has an output. all commented code is working towards persisting but has altered it so that in the create method in UserService, the user is returned w a value of nil. I dont yet know how to fix this